+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2.10 branch
 ### 2.10.1
 * now pages have categories;
+* now the install shell can create user groups;
 * added links on userbar for posts categories and tags;
 * added `userbar()` method for `AppViews`. This simplifies the code to add 
 	elements to the userbar;